N-Queens Playground v0.0.1 NQueens

NQueens is a module that encapsulates an algorithm to solve the N-Queens Problem.

Its only important public function is solve/1 but I’ve left the rest exported for playing with in the REPL.

Types

board :: [[String.t]]

A board is a list of lists of strings that represents an NxN game board, which will be used to show queen placements and open versus unusable spaces. This is not used for calculations, and is entirely meant for display purposes.

branch_state :: {non_neg_integer, [position]}

A branch_state represents the state for a branch of our N-Queens calculation. It is represented as a 2-tuple containing n and positions, where positions is a list of type position.

position :: {non_neg_integer, non_neg_integer}

A position is a 2-tuple representing a row, column pair.

solution :: [position]

A solution is a list of position. This is what our solve function will return.

Functions

blocked_positions(n, used_positions)

Specs

blocked_positions(non_neg_integer, [position]) :: [position]

blocked_positions/2 will return a list of those positions that are blocked on the board, given a list of queen placements.

blocked_positions_for_position(n, arg)

Specs

blocked_positions_for_position(non_neg_integer, position) :: [position]

blocked_positions_for_position/2 returns those positions that are unavailable due to a queen being placed in the given position.

column_for(n, arg)

Specs

column_for(non_neg_integer, position) :: [position]

Given a position for queen placement, column_for/2 returns all the positions that would be unavailable for placement because they are on the queen’s column.

diagonal_for(n, arg)

Specs

diagonal_for(non_neg_integer, position) :: [position]

Given a position for queen placement, diagonal_for/2 returns all the positions that would be unavailable for placement because they are on a diagonal with the queen.

possible_positions(n)

Specs

possible_positions(non_neg_integer) :: [position]

possible_positions/1 returns all possible positions in an NxN grid.

possible_positions(n, used_positions)

Specs

possible_positions(non_neg_integer, [position]) :: [position]

possible_positions/2 returns all possible positions to place a queen, omitting those positions that are blocked by any of the currently placed queens.

render(n, solution)

Specs

render(non_neg_integer, solution) :: board

render/2 returns a 2-dimensional list of size n in each dimension, showing a Q in each position that is present in the solution, and an 0 in each position that is not present in the solution.

row_for(n, arg)

Specs

row_for(non_neg_integer, position) :: [position]

Given a position for queen placement, row_for/2 returns all the positions that would be unavailable for placement because they are on the queen’s row.

solve(n)

Specs

solve(non_neg_integer) :: [solution]

solve/1 takes an N and returns a list of solutions for that N-Queens problem. The solutions take the form {non_neg_integer, non_neg_integer}, which is the same as type position.

iex> NQueens.solve(4)
[[{0, 2}, {1, 0}, {2, 3}, {3, 1}], [{0, 1}, {1, 3}, {2, 0}, {3, 2}]]